feat: add useAuthFirebase & ServerAddress value in store
