+++ src/pages/Welcome/Welcome.tsx
@@ -1,14 +1,13 @@
 import Link from '@/components/Welcome/Links/Link';
 import Team from '@/components/Welcome/Team/Team';
 import styles from './welcome.module.scss';
-import { useAppDispatch, useAppSelector } from '@/store/hooks';
-import isAuthFirebase from '@/utils/auth/isAuth';
+import { useAppSelector } from '@/store/hooks';
+import useAuthFirebase from '@/utils/auth/useAuthFirebase';
 import { selectContentWelcome } from '@/store/slices/languageSlice';
 
 const Welcome = () => {
   const content = useAppSelector(selectContentWelcome);
-  const dispatch = useAppDispatch();
-  isAuthFirebase(dispatch);
+  useAuthFirebase();
 
   return (
     <div className={styles.welcome}>
