+++ src/utils/auth/useAuthFirebase.tsx
@@ -1,0 +1,26 @@
+import { getAuth, onAuthStateChanged, User } from 'firebase/auth';
+import { useNavigate } from 'react-router';
+import { WELCOME_PATH } from '@/utils/const/const';
+import { setAuth } from '@/store/slices/authSlice';
+import { setUserDisplayName } from '@/store/slices/firebaseUserSlice';
+import { useAppDispatch } from '@/store/hooks';
+import { useEffect } from 'react';
+
+const useAuthFirebase = () => {
+  const dispatch = useAppDispatch();
+  const auth = getAuth();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    onAuthStateChanged(auth, (user: User | null) => {
+      if (!user) {
+        navigate(WELCOME_PATH);
+      } else {
+        dispatch(setUserDisplayName(user.displayName!));
+        dispatch(setAuth(true));
+      }
+    });
+  }, [auth, dispatch, navigate]);
+};
+
+export default useAuthFirebase;
